Rename: temporary modify capitalization to uppercase
